requirements.txt
anthropic>=0.40.0
flask>=3.0.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
tabulate>=0.9.0
colorama>=0.4.6
# Note: multitasking may need manual install from source if pip build fails:
# pip download multitasking --no-deps && tar xzf multitasking*.tar.gz && cd multitasking* && python setup.py install
